+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added option `neato -n3` for fit ends of edges to nodes' borders when edge
   trajectory is set. #2736
 
+### Fixed
+
+- Some Undefined Behavior in sparse matrix calculations has been removed.
+
 ## [14.0.5] – 2025-11-28
 
 ### Fixed